+++ src/Entity/Formation.php
@@ -20,13 +20,14 @@ use Doctrine\Common\Collections\Collection;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Event\PreFlushEventArgs;
 use Doctrine\ORM\Mapping as ORM;
-use Gedmo\Sluggable\Handler\RelativeSlugHandler;
 use Symfony\Component\Serializer\Annotation\Groups;
 use Gedmo\Mapping\Annotation as Gedmo;
 
 use Symfony\Component\Serializer\Annotation\Ignore;
 
 #[ORM\Entity(repositoryClass: FormationRepository::class)]
+#[ORM\Index(name: 'sigle_formation', columns: ['sigle'], flags: ['fulltext'])]
+#[ORM\Index(name: 'slug_formation', columns: ['slug'], flags: ['fulltext'])]
 #[ORM\HasLifecycleCallbacks]
 class Formation
 {
@@ -107,6 +108,7 @@ class Formation
 
     #[Ignore]
     #[ORM\OneToMany(mappedBy: 'formation', targetEntity: Parcours::class)]
+    #[ORM\OrderBy(['libelle' => 'ASC'])]
     private Collection $parcours;
 
     #[ORM\OneToMany(mappedBy: 'formation', targetEntity: BlocCompetence::class)]