+++ src/Entity/Semestre.php
@@ -30,7 +30,7 @@ class Semestre
     private ?int $ordre = null;
 
     #[MaxDepth(1)]
-    #[ORM\OneToMany(mappedBy: 'semestre', targetEntity: Ue::class)]
+    #[ORM\OneToMany(mappedBy: 'semestre', targetEntity: Ue::class, fetch: 'EAGER')]
     #[ORM\OrderBy(['ordre' => 'ASC'])]
     private Collection $ues;
 
@@ -45,7 +45,7 @@ class Semestre
     #[ORM\OneToMany(mappedBy: 'semestre', targetEntity: SemestreMutualisable::class)]
     private Collection $semestreMutualisables;
 
-    #[ORM\ManyToOne(inversedBy: 'semestres')]
+    #[ORM\ManyToOne(inversedBy: 'semestres', fetch: 'EAGER')]
     private ?SemestreMutualisable $semestreRaccroche = null;
 
     #[ORM\Column(type: Types::TEXT, nullable: true)]
@@ -53,6 +53,9 @@ class Semestre
 
     #[ORM\Column]
     private ?bool $nonDispense = false;
+
+    #[ORM\Column(length: 5, nullable: true)]
+    private ?string $codeApogee = null;
 
     public function __construct()
     {
@@ -240,6 +243,18 @@ class Semestre
     public function setNonDispense(bool $nonDispense): self
     {
         $this->nonDispense = $nonDispense;
+
+        return $this;
+    }
+
+    public function getCodeApogee(): ?string
+    {
+        return $this->codeApogee;
+    }
+
+    public function setCodeApogee(?string $codeApogee): static
+    {
+        $this->codeApogee = $codeApogee;
 
         return $this;
     }